backend/requirements.txt
tensorflow
numpy
pandas
sklearn
torch
seaborn
matplotlib
Pillow
PyYaml
PillowImage